add  more check when create .env

+++ scripts/create-environment.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
 if [ -d "$HOME/.osmosisd" ]; then
-  echo "OSMOSISD_ENVIRONMENT=mainnet" > $HOME/.osmosisd/.env
+  if [ ! -f $HOME/.osmosisd/.env ]; then
+    echo "OSMOSISD_ENVIRONMENT=mainnet" > $HOME/.osmosisd/.env
+  fi
+else
+  echo "./osmosisd not exist"
 fi
+
+cat $HOME/.osmosisd/.env
